Persistent Menu options. Persistent values class created
