+++ Assets/Scripts/Menu/SoundOptions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5ac1cd619befc414ab9339d38c247a97
+timeCreated: 1479571858
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
