+++ Assets/Scripts/PersistentValues.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 9c3c31beccc424e478a27435123fd981
+timeCreated: 1479572342
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
